PSO SoA rewrite
I realized the original implementation was too inflexible to
idiomatically implement the interleaved LP optimization, so I just
re-implemented everything storing the particles as a matrix.

This allows movement operation with more context information per
particle, that is necessary for the weird mixed optimization for the
supply chains.

I still need to implement the Supply Chain side of this though...